+++ benches/stack.rs
@@ -5,7 +5,6 @@
 // Copyright (c) DUSK NETWORK. All rights reserved.
 
 use criterion::{black_box, criterion_group, criterion_main, Criterion};
-use microkelvin::{HostStore, StoreRef};
 use rusk_vm::{Contract, ContractId, GasMeter, NetworkState};
 use stack::Stack;
 
@@ -38,12 +37,11 @@ fn stack_bench(c: &mut Criterion) {
         ".wasm"
     ));
 
-    let store = StoreRef::new(HostStore::new());
-    let contract = Contract::new(&stack, code.to_vec(), &store);
+    let mut network = NetworkState::new();
 
-    let mut network = NetworkState::new(store);
-
+    let contract = Contract::new(&stack, code.to_vec(), network.store());
     let contract_id = network.deploy(contract).unwrap();
+
     let mut gas = GasMeter::with_limit(1_000_000_000_000);
     c.bench_function("stack 64", |b| {
         b.iter(|| {